Contract names and some cleanup.
