+++ ocean_lib/web3_internal/constants.py
@@ -8,9 +8,7 @@ This module holds following default values and constants.
 
 """
 
-# TODO: since this module doesn't hold gas-related constants, maybe we should move these
 ZERO_ADDRESS = "0x0000000000000000000000000000000000000000"
-BLOB = "f8929916089218bdb4aa78c3ecd16633afd44b8aef89299160"
 
 MAX_UINT256 = 2**256 - 1
