+++ tests/integration/remote/test_mumbai.py
@@ -71,7 +71,6 @@ def test_ocean_tx__create_data_nft(tmp_path):
 def _remote_config_mumbai(tmp_path):
     config = {
         "NETWORK_NAME": "mumbai",
-        "TRANSACTION_TIMEOUT": 60,
         "METADATA_CACHE_URI": "https://v4.aquarius.oceanprotocol.com",
         "PROVIDER_URL": "https://v4.provider.mumbai.oceanprotocol.com",
         "DOWNLOADS_PATH": "consume-downloads",
